refac: LottoValidator 삭제 및 검증 로직 이동
